# internal/api/server.go
// Package api exposes Tuck's HTTP interface.
package api

import (
	"context"
	"encoding/json"
	"errors"
	"net/http"

	"github.com/NAGenaev/tuck/internal/audit"
	"github.com/NAGenaev/tuck/internal/barrier"
	"github.com/NAGenaev/tuck/internal/core"
	"github.com/NAGenaev/tuck/internal/metrics"
	"github.com/NAGenaev/tuck/internal/ui"
)

const maxBodyBytes = 1 << 20 // 1 MiB

type contextKey int

const tokenCtxKey contextKey = iota

// Server adapts a core.Core to HTTP.
type Server struct {
	core  *core.Core
	audit *audit.Logger
}

// New returns an HTTP server over the given core with a no-op audit logger.
func New(c *core.Core) *Server { return NewWithAudit(c, audit.Nop()) }

// NewWithAudit returns an HTTP server with the given audit logger.
func NewWithAudit(c *core.Core, l *audit.Logger) *Server {
	return &Server{core: c, audit: l}
}

// Handler builds the route table.
func (s *Server) Handler() http.Handler {
	mux := http.NewServeMux()

	// Embedded web dashboard — served at /ui/
	mux.Handle("/ui/", http.StripPrefix("/ui", ui.Handler()))

	mux.HandleFunc("GET /v1/health", s.health)

	// Sys endpoints: seal-status and unseal are unauthenticated; seal requires root.
	mux.HandleFunc("GET /v1/sys/seal-status", s.getSealStatus)
	mux.HandleFunc("GET /v1/sys/ready", s.getReady)
	mux.HandleFunc("POST /v1/sys/unseal", s.postUnseal)
	mux.HandleFunc("POST /v1/sys/seal", s.requireToken(s.postSeal))
	mux.HandleFunc("GET /v1/sys/snapshot", s.requireToken(s.getSnapshot))
	mux.HandleFunc("POST /v1/sys/rotate", s.requireToken(s.postRotate))

	mux.HandleFunc("GET /metrics", metrics.Handler())

	mux.HandleFunc("GET /v1/secret/{path...}", s.requireToken(s.getSecret))
	mux.HandleFunc("PUT /v1/secret/{path...}", s.requireToken(s.putSecret))
	mux.HandleFunc("DELETE /v1/secret/{path...}", s.requireToken(s.deleteSecret))
	mux.HandleFunc("LIST /v1/secret/{path...}", s.requireToken(s.listSecrets))

	mux.HandleFunc("POST /v1/auth/token", s.requireToken(s.createToken))
	mux.HandleFunc("GET /v1/auth/token/{id}", s.requireToken(s.lookupToken))
	mux.HandleFunc("DELETE /v1/auth/token/{id}", s.requireToken(s.revokeToken))
	mux.HandleFunc("POST /v1/auth/token/{id}/renew", s.requireToken(s.renewToken))
	mux.HandleFunc("LIST /v1/auth/token/", s.requireToken(s.listTokens))

	mux.HandleFunc("PUT /v1/policy/{name}", s.requireToken(s.putPolicy))
	mux.HandleFunc("GET /v1/policy/{name}", s.requireToken(s.getPolicy))
	mux.HandleFunc("DELETE /v1/policy/{name}", s.requireToken(s.deletePolicy))
	mux.HandleFunc("LIST /v1/policy/", s.requireToken(s.listPolicies))

	mux.HandleFunc("POST /v1/auth/kubernetes/login", s.loginK8s)
	mux.HandleFunc("PUT /v1/auth/kubernetes/role/{namespace}/{sa}", s.requireToken(s.putK8sRole))
	mux.HandleFunc("GET /v1/auth/kubernetes/role/{namespace}/{sa}", s.requireToken(s.getK8sRole))
	mux.HandleFunc("DELETE /v1/auth/kubernetes/role/{namespace}/{sa}", s.requireToken(s.deleteK8sRole))

	// KV v2 — versioned secrets
	mux.HandleFunc("PUT /v2/secret/{path...}", s.requireToken(s.v2WriteSecret))
	mux.HandleFunc("GET /v2/secret/{path...}", s.requireToken(s.v2ReadSecret))
	mux.HandleFunc("DELETE /v2/secret/{path...}", s.requireToken(s.v2DeleteSecret))
	mux.HandleFunc("LIST /v2/secret/{path...}", s.requireToken(s.v2ListSecrets))
	mux.HandleFunc("POST /v2/secret/undelete/{path...}", s.requireToken(s.v2Undelete))
	mux.HandleFunc("POST /v2/secret/destroy/{path...}", s.requireToken(s.v2Destroy))
	mux.HandleFunc("GET /v2/secret/metadata/{path...}", s.requireToken(s.v2GetMeta))
	mux.HandleFunc("PUT /v2/secret/metadata/{path...}", s.requireToken(s.v2UpdateMeta))
	mux.HandleFunc("DELETE /v2/secret/metadata/{path...}", s.requireToken(s.v2DeleteMeta))
	mux.HandleFunc("LIST /v2/secret/metadata/{path...}", s.requireToken(s.v2ListMeta))

	// OpenAPI spec
	mux.HandleFunc("GET /openapi.json", serveOpenAPI)

	return audit.Middleware(s.audit, mux)
}

// requireToken extracts and validates X-Tuck-Token, then stores the token ID
// in context. Returns 401 on missing/invalid token, 503 if the barrier is sealed.
func (s *Server) requireToken(next http.HandlerFunc) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		id := r.Header.Get("X-Tuck-Token")
		if id == "" {
			writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "missing X-Tuck-Token header"})
			return
		}
		if _, err := s.core.Authenticate(r.Context(), id); err != nil {
			if errors.Is(err, barrier.ErrSealed) {
				writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "sealed"})
				return
			}
			writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "invalid or expired token"})
			return
		}
		next(w, r.WithContext(context.WithValue(r.Context(), tokenCtxKey, id)))
	}
}

func tokenFromCtx(ctx context.Context) string {
	id, _ := ctx.Value(tokenCtxKey).(string)
	return id
}

func (s *Server) health(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, map[string]any{"sealed": s.core.Sealed()})
}

func writeErr(w http.ResponseWriter, err error) {
	switch {
	case errors.Is(err, barrier.ErrSealed):
		writeJSON(w, http.StatusServiceUnavailable, map[string]string{"error": "sealed"})
	case errors.Is(err, core.ErrTokenInvalid):
		writeJSON(w, http.StatusUnauthorized, map[string]string{"error": "invalid token"})
	case errors.Is(err, core.ErrUnauthorized):
		writeJSON(w, http.StatusForbidden, map[string]string{"error": "permission denied"})
	default:
		writeJSON(w, http.StatusInternalServerError, map[string]string{"error": err.Error()})
	}
}

func writeJSON(w http.ResponseWriter, status int, v any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(v)
}